NOJIRA-api-error-response-0b
Move assertErrorResponse from error.go to error_test.go so testing and net/http/httptest are not linked into the production api-manager binary. assertErrorResponse stays in package server so any *_test.go file in the package (including handler tests in later PRs) can still call it.

- bin-api-manager: Move assertErrorResponse test helper out of production error.go into error_test.go
