+++ bin-api-manager/server/error_test.go
@@ -119,3 +119,29 @@ func TestAssertErrorResponseHelper(t *testing.T) {
 
 	assertErrorResponse(t, w, cerrors.StatusInvalidArgument, "INVALID_ID")
 }
+
+// assertErrorResponse is a test helper shared across handler tests in
+// subsequent migration PRs. It asserts the HTTP status code matches
+// the canonical Status AND the response body's status/reason fields
+// match the expected values.
+func assertErrorResponse(t *testing.T, w *httptest.ResponseRecorder, wantStatus cerrors.Status, wantReason string) {
+	t.Helper()
+	if got, want := w.Code, cerrors.HTTPStatusFor(wantStatus); got != want {
+		t.Errorf("status code = %d want %d", got, want)
+	}
+	var body struct {
+		Error struct {
+			Status string `json:"status"`
+			Reason string `json:"reason"`
+		} `json:"error"`
+	}
+	if err := json.Unmarshal(w.Body.Bytes(), &body); err != nil {
+		t.Fatalf("unmarshal body: %v; body=%s", err, w.Body.String())
+	}
+	if body.Error.Status != string(wantStatus) {
+		t.Errorf("status field = %q want %q", body.Error.Status, wantStatus)
+	}
+	if body.Error.Reason != wantReason {
+		t.Errorf("reason = %q want %q", body.Error.Reason, wantReason)
+	}
+}
